feat(pyxld-kris): Clean up mock data util in ideaspace
