--- apps/ideas/src/utils/useMockData.js
@@ -1,832 +1,0 @@
-import axios from 'axios';
-import AxiosMockAdapter from 'axios-mock-adapter';
-import { useUserDataContext }from '@devlaunchers/components/context/UserDataContext';
-
-let isInitialized = false;
-export default function useMockDataInDevelopment() {
-    if (process.env.NEXT_PUBLIC_NAME == "DEVELOPMENT" && !isInitialized)  {
-
-        const axiosMockInstance = axios.create();
-        let mock = new AxiosMockAdapter(axios, { delayResponse: 0 });
-
-        let mockUserData = {
-        "id": 1,
-        "username": "Kris_Gano",
-        "email": "krisgano@devlaunchers.com",
-        "provider": "google",
-        "confirmed": true,
-        "blocked": null,
-        "role": {
-            "id": 1,
-            "name": "Authenticated",
-            "description": "Default role given to authenticated user.",
-            "type": "authenticated"
-        },
-        "point": {
-            "id": 1,
-            "totalPoints": 0,
-            "totalSeasonPoints": 0,
-            "availablePoints": 0,
-            "volunteerHours": 0,
-            "user": 6,
-            "published_at": "2021-07-20T21:54:40.108Z",
-            "created_at": "2021-07-20T21:54:40.151Z",
-            "updated_at": "2022-10-01T07:55:08.367Z",
-            "projectMeetingMinutes": null
-        },
-        "profile": {
-            "id": 1,
-            "displayName": null,
-            "email": null,
-            "bio": "I start things, I run things, I do a lot! Some of the things I do include: building systems, modeling businesses and software, working with video games, leveraging a variety of mediums to tell stories and sell ideas, creating products focused on users, consulting on tech/organizational decisions for businesses, and getting down and dirty with code. In a nutshell, I love making ideas happen and continue to work hard to grow the skills needed to do so!\n\nI've set out to help as many people as I can while building awesome stuff. My life is good.",
-            "profilePictureUrl": "https://lh3.googleusercontent.com/a/AATXAJxZ3nb3nA60iLAv1-RqQizgB871aHjJp4LTwi55=s96-c",
-            "level": null,
-            "user": 6,
-            "published_at": "2021-07-20T21:54:39.784Z",
-            "created_at": "2021-07-20T21:54:39.830Z",
-            "updated_at": "2022-10-03T05:45:16.669Z"
-        },
-        "created_at": "2021-07-20T21:54:40.390Z",
-        "updated_at": "2022-10-01T07:55:08.698Z",
-        "discordId": null,
-        "googleId": "105268478633532816935",
-        "discordAvatar": null,
-        "discordDiscriminator": null,
-        "discordUsername": null,
-        "userId": "9187876b-ca72-4108-9680-169049ee45b4",
-        "birthday": null,
-        "country": null,
-        "zipCode": null,
-        "hasAcceptedTermsOfService": null,
-        "hasSubscribedEmails": null,
-        "interests": [
-            {
-                "id": 17,
-                "published_at": "2021-08-11T05:41:33.064Z",
-                "created_at": "2021-08-11T05:41:28.424Z",
-                "updated_at": "2021-08-11T05:41:33.478Z",
-                "interest": "Node.js",
-                "categories": []
-            },
-            {
-                "id": 7,
-                "published_at": "2021-08-11T05:37:11.747Z",
-                "created_at": "2021-08-11T05:37:07.908Z",
-                "updated_at": "2021-08-11T05:37:11.973Z",
-                "interest": "Web Dev",
-                "categories": []
-            },
-            {
-                "id": 1,
-                "published_at": "2021-08-11T05:35:35.681Z",
-                "created_at": "2021-08-11T05:35:29.489Z",
-                "updated_at": "2021-08-11T05:35:35.960Z",
-                "interest": "Blockchain",
-                "categories": []
-            },
-            {
-                "id": 4,
-                "published_at": "2021-08-11T05:36:30.868Z",
-                "created_at": "2021-08-11T05:36:25.579Z",
-                "updated_at": "2021-08-11T05:36:31.097Z",
-                "interest": "UI/UX",
-                "categories": []
-            },
-            {
-                "id": 6,
-                "published_at": "2021-08-11T05:36:57.160Z",
-                "created_at": "2021-08-11T05:36:54.799Z",
-                "updated_at": "2021-08-11T05:36:57.390Z",
-                "interest": "Marketing",
-                "categories": []
-            },
-            {
-                "id": 5,
-                "published_at": "2021-08-11T05:36:44.987Z",
-                "created_at": "2021-08-11T05:36:41.676Z",
-                "updated_at": "2021-08-11T05:36:45.213Z",
-                "interest": "Visual Art",
-                "categories": []
-            },
-            {
-                "id": 15,
-                "published_at": "2021-08-11T05:40:54.513Z",
-                "created_at": "2021-08-11T05:40:51.568Z",
-                "updated_at": "2021-08-11T08:22:02.188Z",
-                "interest": "JavaScript",
-                "categories": []
-            },
-            {
-                "id": 16,
-                "published_at": "2021-08-11T05:41:14.781Z",
-                "created_at": "2021-08-11T05:41:11.760Z",
-                "updated_at": "2021-08-11T05:41:15.012Z",
-                "interest": "React",
-                "categories": []
-            },
-            {
-                "id": 3,
-                "published_at": "2021-08-11T05:36:12.409Z",
-                "created_at": "2021-08-11T05:36:09.405Z",
-                "updated_at": "2021-08-11T05:36:12.637Z",
-                "interest": "Game Design",
-                "categories": []
-            },
-            {
-                "id": 24,
-                "published_at": "2021-08-11T08:18:16.280Z",
-                "created_at": "2021-08-11T08:18:13.488Z",
-                "updated_at": "2021-08-11T08:18:16.570Z",
-                "interest": "Git",
-                "categories": []
-            },
-            {
-                "id": 2,
-                "published_at": "2021-08-11T05:35:57.994Z",
-                "created_at": "2021-08-11T05:35:54.653Z",
-                "updated_at": "2021-08-11T05:35:58.223Z",
-                "interest": "Game Development",
-                "categories": []
-            },
-            {
-                "id": 27,
-                "published_at": "2021-08-11T08:18:59.797Z",
-                "created_at": "2021-08-11T08:18:56.725Z",
-                "updated_at": "2021-08-11T08:19:00.018Z",
-                "interest": "Web Design",
-                "categories": []
-            }
-        ]
-        };
-
-        let mockIdeaData = [
-            {
-                "id": 195,
-                numLikes: 10,
-                emotes: [
-                    {type: "smile", amount: 2},
-                    {type: "heart", amount: 10},
-                ],
-                "ideaName": "mock data WORKED!",
-                "tagline": "dfg",
-                "description": "dfgdf",
-                "targetAudience": "dfg",
-                "hourCommitmentMin": 0,
-                "hourCommitmentMax": 0,
-                "author": {
-                    id: 1,
-                },
-                "difficultyLevel": "Beginner",
-                "published_at": "2022-07-19T14:07:19.505Z",
-                "created_at": "2022-07-19T14:07:19.606Z",
-                "updated_at": "2022-07-19T14:07:20.782Z",
-                "discord": "ViniciusBonani",
-                "email": "viniciusbonani@artificialmentor.com.br",
-                "calendly": "OK",
-                "features": "dfg",
-                "experience": "gfd",
-                "skills": [
-                    {
-                        "id": 243,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 244,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 99,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": []
-            },
-            {
-                "id": 123,
-                "ideaName": "Idea Tool",
-                "tagline": "",
-                "description": "An online, community based idea or problem solving too",
-                "targetAudience": "",
-                "hourCommitmentMin": 0,
-                "hourCommitmentMax": 0,
-                "author": {
-                    id: 1
-                },
-                "difficultyLevel": "Beginner",
-                "published_at": "2022-05-23T23:06:05.427Z",
-                "created_at": "2022-05-23T23:06:05.467Z",
-                "updated_at": "2022-05-23T23:06:06.036Z",
-                "discord": "Unknown",
-                "email": "unknown@gmail.com",
-                "calendly": "No",
-                "features": "",
-                "experience": "",
-                "skills": [
-                    {
-                        "id": 110,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 111,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 72,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": [
-                    {
-                        "id": 75,
-                        "text": "Great idea!",
-                        "idea_card": 123,
-                        "author": "abc",
-                        "published_at": "2022-09-29T17:14:39.971Z",
-                        "created_at": "2022-09-29T17:14:40.016Z",
-                        "updated_at": "2022-09-29T17:14:40.096Z"
-                    },
-                    {
-                        "id": 52,
-                        "text": "",
-                        "idea_card": 123,
-                        "author": "",
-                        "published_at": "2022-08-02T15:08:11.467Z",
-                        "created_at": "2022-08-02T15:08:11.597Z",
-                        "updated_at": "2022-08-02T15:08:11.728Z"
-                    }
-                ]
-            },
-            {
-                "id": 122,
-                "ideaName": "Dev Launcher's Backpack",
-                "tagline": "This is a tagline.",
-                "description": "A Chrome plugin which easily allows Developers to \"bookmark\" learning resources they find as they're browsing the internet",
-                "targetAudience": "This is a target audience.",
-                "hourCommitmentMin": 5,
-                "hourCommitmentMax": 10,
-                "author": {},
-                "difficultyLevel": "Beginner",
-                "published_at": "2022-05-23T23:04:35.503Z",
-                "created_at": "2022-05-23T23:04:35.547Z",
-                "updated_at": "2022-10-06T13:44:31.461Z",
-                "discord": "Unknown",
-                "email": "unknown@gmail.com",
-                "calendly": "No",
-                "features": "",
-                "experience": "",
-                "skills": [
-                    {
-                        "id": 108,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 109,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 71,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": [
-                    {
-                        "id": 36,
-                        "text": "test",
-                        "idea_card": 122,
-                        "author": "test",
-                        "published_at": "2022-05-25T01:20:37.830Z",
-                        "created_at": "2022-05-25T01:20:37.860Z",
-                        "updated_at": "2022-10-06T13:44:31.427Z"
-                    },
-                    {
-                        "id": 43,
-                        "text": "What a neat idea",
-                        "idea_card": 122,
-                        "author": "Kris",
-                        "published_at": "2022-06-11T17:47:14.051Z",
-                        "created_at": "2022-06-11T17:47:14.083Z",
-                        "updated_at": "2022-10-06T13:44:31.427Z"
-                    }
-                ]
-            },
-            {
-                "id": 202,
-                "ideaName": "KnowledgeRepo",
-                "tagline": "KnowledgeRepo",
-                "description": "This project is split into two parts:\n\n## [PLUGIN] \n(See https://www.figma.com/file/Xe8GeAamM6EtHbwO9gGkZh/KnowledgeRepo-Specs?node-id=0%3A1 for basic plugin UI and schema for backend) A Chrome plugin which easily allows Developers to \"bookmark\" learning resources they find as they're browsing the internet, which would then be added and indexed in the \"Dev Launchers Knowledge Repository\". These resources would also be added per user to their own personal Dev Launchers Knowledge Repository. Learning resources would be indexable by Devs based on: \n         - Intended Audience/Targeted Skill Level\n         - Technology or Category\n         - Medium or Learning Style\n\n---\n\n## [LEARNING TOOL] \nA site or app that allows learners to explore these curated learning resources based on different skills they're interested in, and their preferred learning methods. ",
-                "targetAudience": "Developers",
-                "hourCommitmentMin": 1,
-                "hourCommitmentMax": 2,
-                "author": {
-                    id: 1
-                },
-                "difficultyLevel": "Intermediate",
-                "published_at": "2022-09-06T23:00:27.901Z",
-                "created_at": "2022-09-06T22:59:49.953Z",
-                "updated_at": "2022-09-07T02:38:18.221Z",
-                "discord": null,
-                "email": null,
-                "calendly": null,
-                "features": null,
-                "experience": null,
-                "skills": [],
-                "openPositions": [],
-                "comments": []
-            },
-            {
-                "id": 200,
-                "ideaName": "asfasd",
-                "tagline": "",
-                "description": "b",
-                "targetAudience": "c",
-                "hourCommitmentMin": 0,
-                "hourCommitmentMax": 0,
-                "author": {
-                    id: 2
-                },
-                "difficultyLevel": "Beginner",
-                "published_at": "2022-08-03T21:05:36.512Z",
-                "created_at": "2022-08-03T21:05:36.622Z",
-                "updated_at": "2022-08-03T21:05:37.618Z",
-                "discord": "sadfsd",
-                "email": "abc123@gmail.com",
-                "calendly": "",
-                "features": "d",
-                "experience": "a",
-                "skills": [
-                    {
-                        "id": 249,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 250,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 101,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": []
-            },
-            {
-                "id": 124,
-                "ideaName": "Dev Launcher's Classroom",
-                "tagline": "The best classroom ever!",
-                "description": "Virtual campus or office pages with customizable rooms and structures which other businesses can create, manage, and run their operations from",
-                "targetAudience": "Students",
-                "hourCommitmentMin": 3,
-                "hourCommitmentMax": 4,
-                "author": {
-                    id: 2
-                },
-                "difficultyLevel": "Beginner",
-                "published_at": "2022-05-23T23:07:46.515Z",
-                "created_at": "2022-05-23T23:07:46.553Z",
-                "updated_at": "2022-09-01T21:36:08.116Z",
-                "discord": "Unknown",
-                "email": "unknown@gmail.com",
-                "calendly": "No",
-                "features": "",
-                "experience": "",
-                "skills": [
-                    {
-                        "id": 112,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 113,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 73,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": []
-            },
-            {
-                "id": 196,
-                "ideaName": "abc",
-                "tagline": "This is a tagline.",
-                "description": "This is a description.",
-                "targetAudience": "Our targer audience.",
-                "hourCommitmentMin": 3,
-                "hourCommitmentMax": 5,
-                "author": {
-                    id: 1
-                },
-                "difficultyLevel": "Beginner",
-                "published_at": "2022-08-02T15:09:44.488Z",
-                "created_at": "2022-08-02T15:09:44.548Z",
-                "updated_at": "2022-10-06T13:42:58.554Z",
-                "discord": "abc",
-                "email": "jordanmcousins@gmail.com",
-                "calendly": "No",
-                "features": "",
-                "experience": "",
-                "skills": [
-                    {
-                        "id": 247,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 248,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 100,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": []
-            },
-            {
-                "id": 189,
-                "ideaName": "Test Button Next4",
-                "tagline": "o",
-                "description": "o",
-                "targetAudience": "o",
-                "hourCommitmentMin": 0,
-                "hourCommitmentMax": 0,
-                "author": {
-                    id: 2
-                },
-                "difficultyLevel": "Advanced",
-                "published_at": "2022-07-15T21:09:27.510Z",
-                "created_at": "2022-07-15T21:09:27.771Z",
-                "updated_at": "2022-09-03T18:26:52.134Z",
-                "discord": "ViniciusBonani",
-                "email": "viniciusbonani@artificialmentor.com.br",
-                "calendly": "OK",
-                "features": "o",
-                "experience": "o",
-                "skills": [
-                    {
-                        "id": 236,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 237,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 98,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": [
-                    {
-                        "id": 51,
-                        "text": "",
-                        "idea_card": 189,
-                        "author": "",
-                        "published_at": "2022-07-18T13:51:06.241Z",
-                        "created_at": "2022-07-18T13:51:06.301Z",
-                        "updated_at": "2022-09-03T18:26:52.100Z"
-                    },
-                    {
-                        "id": 49,
-                        "text": "",
-                        "idea_card": 189,
-                        "author": "",
-                        "published_at": "2022-07-18T13:09:45.941Z",
-                        "created_at": "2022-07-18T13:09:46.014Z",
-                        "updated_at": "2022-09-03T18:26:52.100Z"
-                    },
-                    {
-                        "id": 50,
-                        "text": "Teste",
-                        "idea_card": 189,
-                        "author": "Vinny",
-                        "published_at": "2022-07-18T13:09:54.228Z",
-                        "created_at": "2022-07-18T13:09:54.259Z",
-                        "updated_at": "2022-09-03T18:26:52.100Z"
-                    }
-                ]
-            },
-            {
-                "id": 203,
-                "ideaName": "A Kris Test",
-                "tagline": "This is a multiline\nANYTHING ELSE\nanswer",
-                "description": "This is a multiline\nIDEA\nanswer",
-                "targetAudience": "This is a multiline\nHELPFULNESS\nanswer",
-                "hourCommitmentMin": 1,
-                "hourCommitmentMax": 4,
-                "author": {
-                    id: 2
-                },
-                "difficultyLevel": "Beginner",
-                "published_at": "2022-09-07T01:25:52.088Z",
-                "created_at": "2022-09-07T01:25:52.133Z",
-                "updated_at": "2022-10-06T13:43:19.059Z",
-                "discord": "pyxld-kris",
-                "email": "krisgano@devlaunchers.com",
-                "calendly": "no",
-                "features": "This is a multiline\nFEATURES\nanswer",
-                "experience": "This is a multiline\nEXPERIENCE\nanswer",
-                "skills": [
-                    {
-                        "id": 251,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 252,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 102,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": [
-                    {
-                        "id": 71,
-                        "text": "Does this work?",
-                        "idea_card": 203,
-                        "author": "Kris Gano",
-                        "published_at": "2022-09-07T01:27:05.454Z",
-                        "created_at": "2022-09-07T01:27:05.488Z",
-                        "updated_at": "2022-10-06T13:43:19.026Z"
-                    }
-                ]
-            },
-            {
-                "id": 188,
-                "ideaName": "sdfgedsg",
-                "tagline": " asertygsery ",
-                "description": "sdgedsrgesrg",
-                "targetAudience": "wsegfesragrt wert ",
-                "hourCommitmentMin": 5,
-                "hourCommitmentMax": 8,
-                "author": {
-                    id: 1
-                },
-                "difficultyLevel": "Beginner",
-                "published_at": "2022-06-30T23:38:19.018Z",
-                "created_at": "2022-06-30T23:38:19.058Z",
-                "updated_at": "2022-10-06T13:43:44.141Z",
-                "discord": "dsfghdfsghrtjh",
-                "email": "desrg@sg.desgf",
-                "calendly": "asdfgdsges",
-                "features": " aerygesrat",
-                "experience": "sfewwesfwsef",
-                "skills": [
-                    {
-                        "id": 225,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 226,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 97,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": []
-            },
-            {
-                "id": 125,
-                "ideaName": "Ergonomic Digits",
-                "tagline": "Digits that are ergonomic!",
-                "description": "Create an ergonomic layout for mobile screens with the goal of minimizing thumb movement.",
-                "targetAudience": "People with thumbs",
-                "hourCommitmentMin": 2,
-                "hourCommitmentMax": 4,
-                "author": {
-                    id: 2
-                },
-                "difficultyLevel": "Intermediate",
-                "published_at": "2022-05-23T23:15:05.138Z",
-                "created_at": "2022-05-23T23:15:05.188Z",
-                "updated_at": "2022-10-06T13:44:00.005Z",
-                "discord": "Unknown",
-                "email": "unknown@gmail.com",
-                "calendly": "No",
-                "features": "",
-                "experience": "",
-                "skills": [
-                    {
-                        "id": 114,
-                        "skill": "Web Developer"
-                    },
-                    {
-                        "id": 115,
-                        "skill": "AI / ML"
-                    }
-                ],
-                "openPositions": [
-                    {
-                        "id": 74,
-                        "title": "Developer, UX/UX",
-                        "description": "Lots of programing",
-                        "isHidden": false
-                    }
-                ],
-                "comments": [
-                    {
-                        "id": 83,
-                        "text": "",
-                        "idea_card": 125,
-                        "author": "",
-                        "published_at": "2022-11-09T00:03:43.893Z",
-                        "created_at": "2022-11-09T00:03:43.927Z",
-                        "updated_at": "2022-11-09T00:03:43.994Z"
-                    },
-                    {
-                        "id": 79,
-                        "text": "",
-                        "idea_card": 125,
-                        "author": "",
-                        "published_at": "2022-11-09T00:02:23.404Z",
-                        "created_at": "2022-11-09T00:02:23.434Z",
-                        "updated_at": "2022-11-09T00:02:23.492Z"
-                    },
-                    {
-                        "id": 91,
-                        "text": "1111111111111111111111111111111",
-                        "idea_card": 125,
-                        "author": "1111111",
-                        "published_at": "2022-11-15T23:28:10.109Z",
-                        "created_at": "2022-11-15T23:28:10.139Z",
-                        "updated_at": "2022-11-15T23:28:10.202Z"
-                    },
-                    {
-                        "id": 84,
-                        "text": "less than 20",
-                        "idea_card": 125,
-                        "author": "will",
-                        "published_at": "2022-11-09T00:08:12.110Z",
-                        "created_at": "2022-11-09T00:08:12.142Z",
-                        "updated_at": "2022-11-09T00:08:12.209Z"
-                    },
-                    {
-                        "id": 90,
-                        "text": "11111111111111111111111111",
-                        "idea_card": 125,
-                        "author": "1111111",
-                        "published_at": "2022-11-15T23:26:52.727Z",
-                        "created_at": "2022-11-15T23:26:52.760Z",
-                        "updated_at": "2022-11-15T23:26:52.826Z"
-                    },
-                    {
-                        "id": 85,
-                        "text": "characters less than 25",
-                        "idea_card": 125,
-                        "author": "will",
-                        "published_at": "2022-11-09T00:12:19.008Z",
-                        "created_at": "2022-11-09T00:12:19.043Z",
-                        "updated_at": "2022-11-09T00:12:19.111Z"
-                    },
-                    {
-                        "id": 78,
-                        "text": "",
-                        "idea_card": 125,
-                        "author": "",
-                        "published_at": "2022-11-09T00:02:21.144Z",
-                        "created_at": "2022-11-09T00:02:21.209Z",
-                        "updated_at": "2022-11-09T00:02:21.271Z"
-                    },
-                    {
-                        "id": 89,
-                        "text": "111111111111111111111111111111",
-                        "idea_card": 125,
-                        "author": "1111",
-                        "published_at": "2022-11-15T23:26:04.221Z",
-                        "created_at": "2022-11-15T23:26:04.255Z",
-                        "updated_at": "2022-11-15T23:26:04.320Z"
-                    },
-                    {
-                        "id": 87,
-                        "text": "1211111111111111111111111111111",
-                        "idea_card": 125,
-                        "author": "123",
-                        "published_at": "2022-11-15T22:48:23.638Z",
-                        "created_at": "2022-11-15T22:48:23.678Z",
-                        "updated_at": "2022-11-15T22:48:23.744Z"
-                    },
-                    {
-                        "id": 88,
-                        "text": "",
-                        "idea_card": 125,
-                        "author": "",
-                        "published_at": "2022-11-15T23:22:03.665Z",
-                        "created_at": "2022-11-15T23:22:03.759Z",
-                        "updated_at": "2022-11-15T23:22:03.933Z"
-                    },
-                    {
-                        "id": 81,
-                        "text": "",
-                        "idea_card": 125,
-                        "author": "Will",
-                        "published_at": "2022-11-09T00:02:54.984Z",
-                        "created_at": "2022-11-09T00:02:55.014Z",
-                        "updated_at": "2022-11-09T00:02:55.073Z"
-                    },
-                    {
-                        "id": 76,
-                        "text": "Here's a new comment!",
-                        "idea_card": 125,
-                        "author": "Kris",
-                        "published_at": "2022-10-18T07:40:06.278Z",
-                        "created_at": "2022-10-18T07:40:06.319Z",
-                        "updated_at": "2022-10-18T07:40:06.412Z"
-                    },
-                    {
-                        "id": 92,
-                        "text": "111111111111111111111111111111111111",
-                        "idea_card": 125,
-                        "author": "11111111111111111111",
-                        "published_at": "2022-11-15T23:28:22.450Z",
-                        "created_at": "2022-11-15T23:28:22.483Z",
-                        "updated_at": "2022-11-15T23:28:22.549Z"
-                    },
-                    {
-                        "id": 80,
-                        "text": "",
-                        "idea_card": 125,
-                        "author": "",
-                        "published_at": "2022-11-09T00:02:26.706Z",
-                        "created_at": "2022-11-09T00:02:26.736Z",
-                        "updated_at": "2022-11-09T00:02:26.795Z"
-                    },
-                    {
-                        "id": 44,
-                        "text": "Comment test",
-                        "idea_card": 125,
-                        "author": "Vinny",
-                        "published_at": "2022-06-17T16:05:09.843Z",
-                        "created_at": "2022-06-17T16:05:09.923Z",
-                        "updated_at": "2022-10-06T13:43:59.972Z"
-                    },
-                    {
-                        "id": 86,
-                        "text": "less 25",
-                        "idea_card": 125,
-                        "author": "Will",
-                        "published_at": "2022-11-09T00:18:49.680Z",
-                        "created_at": "2022-11-09T00:18:49.713Z",
-                        "updated_at": "2022-11-09T00:18:49.781Z"
-                    },
-                    {
-                        "id": 82,
-                        "text": "yes, let's do it",
-                        "idea_card": 125,
-                        "author": "",
-                        "published_at": "2022-11-09T00:03:17.515Z",
-                        "created_at": "2022-11-09T00:03:17.545Z",
-                        "updated_at": "2022-11-09T00:03:17.605Z"
-                    }
-                ]
-            }
-        ];
-
-        mock
-        .onGet(`${process.env.NEXT_PUBLIC_STRAPI_URL}/idea-cards`)
-        .reply(() => {
-            return [200, mockIdeaData];
-        });
-
-        isInitialized = true;
-    }
-}
-
-let ideaCardsData = {};
